docs/cypher-ingestion/rl_controller.py
import numpy as np
from dataclasses import dataclass
from typing import List, Dict, Tuple
import torch
import torch.nn as nn
import torch.optim as optim
from collections import deque
import random
from neo4j import GraphDatabase
import logging

@dataclass
class Experience:
    state: np.ndarray
    action: int
    reward: float
    next_state: np.ndarray
    done: bool

class QuantumPolicyNetwork(nn.Module):
    def __init__(self, state_size: int, action_size: int):
        super().__init__()
        self.quantum_layer1 = nn.Linear(state_size, 128)
        self.quantum_layer2 = nn.Linear(128, 64)
        self.action_head = nn.Linear(64, action_size)
        self.value_head = nn.Linear(64, 1)
        
    def forward(self, x):
        x = torch.relu(self.quantum_layer1(x))
        x = torch.relu(self.quantum_layer2(x))
        action_probs = torch.softmax(self.action_head(x), dim=-1)
        state_value = self.value_head(x)
        return action_probs, state_value

class PatternEvolutionRL:
    def __init__(self, uri: str, user: str, password: str):
        self.driver = GraphDatabase.driver(uri, auth=(user, password))
        self.logger = logging.getLogger(__name__)
        
        # RL parameters
        self.state_size = 64
        self.action_size = 16
        self.memory = deque(maxlen=10000)
        self.gamma = 0.95
        self.epsilon = 0.1
        self.epsilon_decay = 0.995
        self.learning_rate = 0.001
        
        # Initialize networks
        self.policy_net = QuantumPolicyNetwork(self.state_size, self.action_size)
        self.target_net = QuantumPolicyNetwork(self.state_size, self.action_size)
        self.target_net.load_state_dict(self.policy_net.state_dict())
        
        self.optimizer = optim.Adam(self.policy_net.parameters(), lr=self.learning_rate)
        
    def get_state(self) -> np.ndarray:
        """Get current system state from Neo4j"""
        with self.driver.session() as session:
            result = session.run("""
                MATCH (nm:NeuralMesh)-[:HARMONIZES_WITH]->(cw:ConsciousnessWeave)
                MATCH (qb:QuantumBridge)-[:SYNCHRONIZES_WITH]->(nm)
                RETURN nm.pattern_synthesis as pattern,
                       qb.coherence_level as coherence,
                       cw.neural_harmonics as harmonics
            """)
            data = result.single()
            
            # Convert to feature vector
            state = np.zeros(self.state_size)
            state[0] = float(data['coherence'])
            # Add more feature engineering here
            return state
            
    def select_action(self, state: np.ndarray) -> int:
        """Select action using epsilon-greedy policy"""
        if random.random() < self.epsilon:
            return random.randint(0, self.action_size - 1)
            
        with torch.no_grad():
            state_tensor = torch.FloatTensor(state).unsqueeze(0)
            action_probs, _ = self.policy_net(state_tensor)
            return torch.argmax(action_probs).item()
            
    def execute_action(self, action: int) -> Tuple[float, bool]:
        """Execute selected action in Neo4j and get reward"""
        with self.driver.session() as session:
            # Execute pattern evolution based on action
            result = session.run("""
                MATCH (nm:NeuralMesh {substrate: 'QUANTUM_FIELD'})
                SET nm.last_action = $action
                WITH nm
                MATCH (nm)-[:EVOLVES_THROUGH]->(tn:TemporalNexus)
                SET tn.evolution_step = tn.evolution_step + 1
                RETURN nm.coherence as new_coherence
            """, action=action)
            
            new_coherence = float(result.single()['new_coherence'])
            reward = self._calculate_reward(new_coherence)
            done = new_coherence < 0.3 or new_coherence > 0.95
            
            return reward, done
            
    def _calculate_reward(self, coherence: float) -> float:
        """Calculate reward based on coherence and stability"""
        base_reward = coherence * 2 - 1  # Maps [0,1] to [-1,1]
        stability_bonus = 0.5 if 0.4 <= coherence <= 0.8 else 0
        return base_reward + stability_bonus
        
    def update_policy(self, batch_size: int = 32):
        """Update policy network using experience replay"""
        if len(self.memory) < batch_size:
            return
            
        # Sample batch from memory
        batch = random.sample(self.memory, batch_size)
        states = torch.FloatTensor([exp.state for exp in batch])
        actions = torch.LongTensor([exp.action for exp in batch])
        rewards = torch.FloatTensor([exp.reward for exp in batch])
        next_states = torch.FloatTensor([exp.next_state for exp in batch])
        dones = torch.FloatTensor([exp.done for exp in batch])
        
        # Get current Q values
        current_q_values, _ = self.policy_net(states)
        current_q_values = current_q_values.gather(1, actions.unsqueeze(1))
        
        # Get next Q values
        with torch.no_grad():
            next_q_values, _ = self.target_net(next_states)
            next_q_values = next_q_values.max(1)[0]
            
        # Calculate target Q values
        target_q_values = rewards + (1 - dones) * self.gamma * next_q_values
        
        # Calculate loss and update
        loss = nn.MSELoss()(current_q_values.squeeze(), target_q_values)
        
        self.optimizer.zero_grad()
        loss.backward()
        self.optimizer.step()
        
        # Update target network
        if len(self.memory) % 100 == 0:
            self.target_net.load_state_dict(self.policy_net.state_dict())
            
    async def train(self, episodes: int = 1000):
        """Train the RL agent"""
        for episode in range(episodes):
            state = self.get_state()
            total_reward = 0
            done = False
            
            while not done:
                # Select and execute action
                action = self.select_action(state)
                reward, done = self.execute_action(action)
                next_state = self.get_state()
                
                # Store experience
                exp = Experience(state, action, reward, next_state, done)
                self.memory.append(exp)
                
                # Update policy
                self.update_policy()
                
                total_reward += reward
                state = next_state
                
                # Decay epsilon
                self.epsilon *= self.epsilon_decay
                
            self.logger.info(f"Episode {episode}: Total Reward = {total_reward}")
            
    def save_model(self, path: str):
        """Save policy network"""
        torch.save(self.policy_net.state_dict(), path)
        
    def load_model(self, path: str):
        """Load policy network"""
        self.policy_net.load_state_dict(torch.load(path))
        self.target_net.load_state_dict(self.policy_net.state_dict())
        
    def close(self):
        """Close database connection"""
        self.driver.close()

if __name__ == "__main__":
    import asyncio
    
    async def main():
        controller = PatternEvolutionRL(
            uri="bolt://localhost:7687",
            user="neo4j",
            password="your-password-here"
        )
        
        try:
            await controller.train(episodes=1000)
            controller.save_model("pattern_evolution_model.pth")
        finally:
            controller.close()
            
    asyncio.run(main())