docs/cypher-ingestion/requirements.txt
neo4j==5.14.1
pytest==8.0.0
typing-extensions==4.9.0
python-dotenv==1.0.0
